fix: kafka
