+++ mq/kafka/kafka_test.go
@@ -1,0 +1,7 @@
+package kafka
+
+import "testing"
+
+func TestProducer(t *testing.T) {
+
+}
